pyproject.toml
[project]
name = "agent-interception"
version = "0.1.0"
description = "Transparent HTTP reverse proxy for intercepting and logging LLM agent interactions"
requires-python = ">=3.11"
dependencies = [
    "starlette>=0.41,<1",
    "uvicorn[standard]>=0.32,<1",
    "httpx>=0.28,<1",
    "aiosqlite>=0.20,<1",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "rich>=13.9,<14",
    "click>=8.1,<9",
    "claude-agent-sdk>=0.1.68,<1",
    "plotly>=5.24,<6",
    "anthropic>=0.84.0",
]

[project.optional-dependencies]
chemcrow = [
    "rdkit>=2024.3,<2026",
    "claude-agent-sdk>=0.1.68,<1",
    "httpx>=0.28,<1",
]
sweagent = [
    "claude-agent-sdk>=0.1.68,<1",
    "python-dotenv>=1.0,<2",
]
toolformer = [
    "claude-agent-sdk>=0.1.68,<1",
    "python-dotenv>=1.0,<2",
    "simpleeval>=1.0,<2",
]

[project.scripts]
agent-interceptor = "agent_interception.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_interception"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.pyright]
pythonVersion = "3.11"
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"

[dependency-groups]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.28,<1",
    "pyright>=1.1,<2",
    "ruff>=0.8,<1",
]
viz-static = [
    "kaleido>=0.2,<1",
]
benchmark = [
    "opentelemetry-api>=1.27,<2",
    "opentelemetry-sdk>=1.27,<2",
    "sumy>=0.11,<1",
    "nltk>=3.9,<4",
    "beautifulsoup4>=4.12,<5",
    "trafilatura>=1.12,<2",
    "ddgs>=9.0,<10",
    "matplotlib>=3.9,<4",
    "pandas>=2.2,<3",
    "python-dotenv>=1.0,<2",
]
